% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainLadder.R
\name{chainLadder}
\alias{chainLadder}
\title{Classical chain-ladder model.}
\usage{
chainLadder(triangle, is_cumulatif = FALSE)
}
\arguments{
\item{triangle}{The run-off triangle. The rows are the reporting years (or occurrence years), the columns the development years.}

\item{is_cumulatif}{Logical. Is the input triangle a cumulative run-off triangle? If `FALSE`, the cumulutive run-off triangle is constructed using the incremental input run-off triangle.}
}
\description{
This function implements the classic chain ladder method on a cumulative run-off triangle.
}
