% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_triangle.R
\name{construct_triangle}
\alias{construct_triangle}
\title{Groups and aggregates the portfolio by e.g. reporting year and development year since reporting}
\usage{
construct_triangle(data, group.var1, group.var2, value, cumulative = FALSE)
}
\arguments{
\item{data}{The portfolio of claims. The portfolio should have one record per development year for each claim.}

\item{group.var1}{Character. Name of the first grouping variable.}

\item{group.var2}{Character. Name of the second grouping variable.}

\item{value}{The name of the variable that will be aggregated over the two grouping variables.}

\item{cumulative}{Logical. Should the cumulative run-off triangle be returned?}
}
\description{
Groups and aggregates the portfolio by e.g. reporting year and development year since reporting
}
