% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrm.R
\name{hirem}
\alias{hirem}
\title{Initiates the construction of a hierarchical reserving model composed of a sequence of layers}
\usage{
hirem(data)
}
\arguments{
\item{data}{The input data for constructing, evaluating and testing the hierarchical reserving model}
}
\description{
Initiates the construction of a hierarchical reserving model composed of a sequence of layers
}
