% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer_cann}
\alias{layer_cann}
\title{Layer estimated using a Combined Actuarial Neural Network (CANN)}
\usage{
layer_cann(
  obj,
  name,
  distribution = "gaussian",
  hidden = c(10, 20, 10),
  dropout.hidden = rep(0.01, 3),
  activation.hidden = rep("tanh", 3),
  activation.output = "linear",
  activation.output.cann = "linear",
  fixed.cann = TRUE,
  loss = "mse",
  optimizer = "nadam",
  epochs = 20,
  batch_size = 1000,
  validation_split = 0.2,
  metrics = NULL,
  family_for_glm = NULL,
  filter = NULL,
  transformation = NULL
)
}
\arguments{
\item{obj}{The hierarchical reserving model}

\item{name}{Character, name of the layer. This name should match the variable name in the data set}

\item{distribution}{Default is tweedie,}

\item{filter}{Function with \itemize{
 \item input: Data set with same structure as the data passed to \code{\link{hirem}}
 \item output: TRUE/FALSE vector with same length as the number of rows in the input data set.\cr
       FALSE indicates that this layer is zero for the current record.
}}

\item{transformation}{Object of class \code{\link{hirem_transformation}} specifying the transformation
applied before modelling this layer.}
}
\description{
Adds a new layer to the hierarchical reserving model. This layer will be estimated using a CANN architecture,
see Schelldorfer, J., & Wuthrich, M. (2019). Nesting Classical Actuarial Models into Neural Networks. Applied Computing eJournal.
}
