% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer_gbm}
\alias{layer_gbm}
\title{Layer estimated using a gradient boosting model}
\usage{
layer_gbm(
  obj,
  name,
  distribution,
  n.trees = 500,
  interaction.depth = 2,
  n.minobsinnode = 10,
  shrinkage = 0.1,
  bag.fraction = 0.5,
  cv.folds = 0,
  select_trees = "perf",
  filter = NULL,
  transformation = NULL
)
}
\arguments{
\item{obj}{The hierarchical reserving model}

\item{name}{Character, name of the layer. This name should match the variable name in the data set}

\item{distribution}{distribution argument passed to \code{\link[gbm]{gbm}}}

\item{n.trees}{n.trees argument passed to \code{\link[gbm]{gbm}}}

\item{interaction.depth}{interaction.depth argument passed to \code{\link[gbm]{gbm}}}

\item{n.minobsinnode}{n.minobsinnode argument passed to \code{\link[gbm]{gbm}}}

\item{shrinkage}{shrinkage argument passed to \code{\link[gbm]{gbm}}}

\item{bag.fraction}{bag.fraction argument passed to \code{\link[gbm]{gbm}}}

\item{select_trees}{Character string specifying the method for selecting the optimal number of trees after fitting the gbm \itemize{
   \item "fixed": Use the number of trees specified in n.trees
   \item "perf": Update the number of trees using \code{\link[gbm]{gbm.perf}}
}}

\item{filter}{Function with \itemize{
 \item input: Data set with same structure as the data passed to \code{\link{hirem}}
 \item output: TRUE/FALSE vector with same length as the number of rows in the input data set.\cr
       FALSE indicates that this layer is zero for the current record.
}}

\item{transformation}{Object of class \code{\link{hirem_transformation}} specifying the transformation
applied before modelling this layer.}
}
\description{
Adds a new layer to the hierarchical reserving model. This layer will be estimated using the \code{\link[gbm]{gbm}} package.
}
