% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer_glm}
\alias{layer_glm}
\title{Layer estimated using a generalized linear model}
\usage{
layer_glm(obj, name, family, filter = NULL, transformation = NULL)
}
\arguments{
\item{obj}{The hierarchical reserving model}

\item{name}{Character, name of the layer. This name should match the variable name in the data set}

\item{family}{family argument passed to the \code{\link[stats]{glm}} function}

\item{filter}{Function with \itemize{
 \item input: Data set with same structure as the data passed to \code{\link{hirem}}
 \item output: TRUE/FALSE vector with same length as the number of rows in the input data set.\cr
       FALSE indicates that this layer is zero for the current record.
}}

\item{transformation}{Object of class \code{\link{hirem_transformation}} specifying the transformation
applied before modelling this layer.}
}
\description{
Adds a new layer to the hierarchical reserving model. This layer will be estimated using the glm framework.
}
