% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer_mlp}
\alias{layer_mlp}
\title{Layer estimated using a multi-layer perceptron model (h2o)}
\usage{
layer_mlp(
  obj,
  name,
  distribution = "tweedie",
  hidden = c(10, 10),
  epochs = 1000,
  train_samples_per_iteration = -1,
  reproducible = T,
  activation = "Tanh",
  nfolds = NULL,
  single_node_mode = FALSE,
  balance_classes = FALSE,
  force_load_balance = FALSE,
  seed = NULL,
  tweedie_power = 1.5,
  score_training_samples = 0,
  score_validation_samples = 0,
  input_dropout_ratio = 0.1,
  hidden_dropout_ratios = 0.5,
  stopping_rounds = 0,
  filter = NULL,
  transformation = NULL
)
}
\arguments{
\item{obj}{The hierarchical reserving model}

\item{name}{Character, name of the layer. This name should match the variable name in the data set}

\item{distribution}{Default is tweedie,}

\item{hidden}{Default is c(10,10)}

\item{epochs}{Default is 1000}

\item{train_samples_per_iteration}{Default is -1}

\item{reproducible}{Default is True,}

\item{activation}{Default is "Tanh",}

\item{single_node_mode}{Default is FALSE,}

\item{balance_classes}{Default is FALSE,}

\item{force_load_balance}{Default is FALSE,}

\item{seed}{Default is 23123,}

\item{tweedie_power}{Default is 1.5,}

\item{score_training_samples}{Default is 0,}

\item{score_validation_samples}{Default is 0,}

\item{input_dropout_ratio}{Default is 0.1}

\item{stopping_rounds}{Default is 0}

\item{filter}{Function with \itemize{
 \item input: Data set with same structure as the data passed to \code{\link{hirem}}
 \item output: TRUE/FALSE vector with same length as the number of rows in the input data set.\cr
       FALSE indicates that this layer is zero for the current record.
}}

\item{transformation}{Object of class \code{\link{hirem_transformation}} specifying the transformation
applied before modelling this layer.}
}
\description{
Adds a new layer to the hierarchical reserving model. This layer will be estimated using the \code{deeplearning} function of the \code{\link[h2o]{h2o}} package.
}
