% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer_mlp_h2o}
\alias{layer_mlp_h2o}
\title{Layer estimated using a multi-layer perceptron model (h2o)}
\usage{
layer_mlp_h2o(
  obj,
  name,
  distribution = "gaussian",
  hidden = c(10, 10),
  epochs = 1000,
  train_samples_per_iteration = -1,
  reproducible = T,
  activation = "Tanh",
  nfolds = NULL,
  single_node_mode = FALSE,
  balance_classes = FALSE,
  force_load_balance = FALSE,
  seed = NULL,
  tweedie_power = 1.5,
  score_training_samples = 0,
  score_validation_samples = 0,
  input_dropout_ratio = 0.1,
  hidden_dropout_ratios = 0.5,
  stopping_rounds = 0,
  filter = NULL,
  transformation = NULL
)
}
\arguments{
\item{obj}{The hierarchical reserving model}

\item{name}{Character, name of the layer. This name should match the variable name in the data set}

\item{distribution}{The distribution used for the simulation. Default is gaussian.}

\item{hidden}{The hidden layer architecture passed to \code{h2o}. Default is c(10,10).}

\item{epochs}{The epochs argument passed to \code{h2o}. Default is 1000}

\item{train_samples_per_iteration}{The train_samples_per_iteration argument passed to \code{h2o}. Default is -1}

\item{reproducible}{The reproducible argument passed to \code{h2o}. Default is True.}

\item{activation}{The activation function (same for all layers) passed to \code{h2o}. Default is "Tanh".}

\item{single_node_mode}{The single_node_mode argument passed to \code{h2o}. Default is FALSE,}

\item{balance_classes}{The balance_classes argument passed to \code{h2o}. Default is FALSE,}

\item{force_load_balance}{The force_load_balance argument passed to \code{h2o}. Default is FALSE,}

\item{seed}{The seed argument passed to \code{h2o}.}

\item{tweedie_power}{The tweedie_power argument passed to \code{h2o}. Only used if distribution is "tweedie".}

\item{score_training_samples}{The score_training_samples argument passed to \code{h2o}. Default is 0,}

\item{score_validation_samples}{The score_validation_samples argument passed to \code{h2o}. Default is 0,}

\item{input_dropout_ratio}{The input_dropout_ratio argument passed to \code{h2o}. Default is 0.1}

\item{hidden_dropout_ratios}{The input_dropout_ratio argument passed to \code{h2o}. Default is 0.5}

\item{stopping_rounds}{The stopping_rounds argument passed to \code{h2o}. Default is 0}

\item{filter}{Function with \itemize{
 \item input: Data set with same structure as the data passed to \code{\link{hirem}}
 \item output: TRUE/FALSE vector with same length as the number of rows in the input data set.\cr
       FALSE indicates that this layer is zero for the current record.
}}

\item{transformation}{Object of class \code{\link{hirem_transformation}} specifying the transformation
applied before modelling this layer.}
}
\description{
Adds a new layer to the hierarchical reserving model. This layer will be estimated using the \code{deeplearning} function of the \code{\link[h2o]{h2o}} package.
}
