% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer_mlp_keras}
\alias{layer_mlp_keras}
\title{Layer estimated using a multi-layer perceptron model with Keras}
\usage{
layer_mlp_keras(
  obj,
  name,
  distribution = "gaussian",
  hidden = c(10, 20, 10),
  dropout = rep(0.01, 3),
  activation = rep("tanh", 3),
  loss = "mse",
  optimizer = "adam",
  epochs = 20,
  batch_size = 1000,
  metrics = NULL,
  filter = NULL,
  transformation = NULL
)
}
\arguments{
\item{obj}{The hierarchical reserving model}

\item{name}{Character, name of the layer. This name should match the variable name in the data set}

\item{distribution}{Default is tweedie,}

\item{filter}{Function with \itemize{
 \item input: Data set with same structure as the data passed to \code{\link{hirem}}
 \item output: TRUE/FALSE vector with same length as the number of rows in the input data set.\cr
       FALSE indicates that this layer is zero for the current record.
}}

\item{transformation}{Object of class \code{\link{hirem_transformation}} specifying the transformation
applied before modelling this layer.}
}
\description{
Adds a new layer to the hierarchical reserving model. This layer will be estimated using the \code{\link[keras]{keras}} package.
}
