% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependence_plots.R
\name{par_dep}
\alias{par_dep}
\title{Marginal effects of a predictor on a hierarchical model layer.}
\usage{
par_dep(object, data, grid)
}
\arguments{
\item{object}{The hierarchical reserving model layer}

\item{data}{The input data set}

\item{grid}{A data frame with the values of the predictor for which marginal effects are to be computed}
}
\description{
This function can be used to obtain the marginal effects of a predictor on the layer of a hierarchical model.
Compatible with layers defined as a GLM, GBM, XGB, DNN or CANN model.
The returned data frame can serve as input to draw partial dependence plots (see examples in the package repository).
}
