% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{register_updater}
\alias{register_updater}
\title{Register a new update function in the hierarchical reserving model}
\usage{
register_updater(obj, update, after = "end")
}
\arguments{
\item{obj}{The hierarchical reserving model}

\item{update}{Function that takes as input a reserving data set and updates stochastic covariates that are not directly modelled in the hirem.}

\item{after}{Name of the layer after which this update function should be applied}
}
\description{
Register an update function that will be called in the simulation algorithm to update the covariates that are not directly modelled in the hierarchical reserving model.
}
