% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.hirem}
\alias{simulate.hirem}
\title{Simulate the future development of claims}
\usage{
\method{simulate}{hirem}(obj, nsim, filter, data, balance.correction = FALSE)
}
\arguments{
\item{obj}{The hierarchical reserving model}

\item{nsim}{Number of paths to simulate for each claim}

\item{filter}{Function with \itemize{
    \item input: Data set with reserving data
    \item output: Subset of records for which the simulation should continue
}}

\item{data}{Last observed record for each claim. From these records onwards the future development is simulated.}

\item{balance.correction}{Logical. Apply a bias correction step: yes or no?}
}
\value{
A data set with the same structure as the input data set, containing multiple simulations for the future development of each claim.
}
\description{
Simulates multiple paths for the future development of each claim
}
