% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_scenario_generator.R
\name{simulate_scenario_baseline}
\alias{simulate_scenario_baseline}
\alias{simulate_scenario_claim_mix}
\alias{simulate_scenario_extreme_event}
\alias{simulate_scenario_change_in_settlement}
\title{Simulate portfolios along various scenarios using the simulation engine}
\usage{
simulate_scenario_baseline(
  seed,
  n = 125000,
  prob.Type = c(0.6, 0.25, 0.15),
  prob.Hidden = c(0.35, 0.45, 0.2),
  max_rep_delay = 2,
  max_set_delay = 20
)

simulate_scenario_claim_mix(
  seed,
  n = 125000,
  prob.Type = c(0.6, 0.25, 0.15),
  prob.Hidden = c(0.35, 0.45, 0.2),
  max_rep_delay = 2,
  max_set_delay = 20,
  inflation = c(-0.02, 0.005, 0.015)
)

simulate_scenario_extreme_event(
  seed,
  n = 125000,
  prob.Type = c(0.6, 0.25, 0.15),
  prob.Hidden = c(0.35, 0.45, 0.2),
  max_rep_delay = 2,
  max_set_delay = 20,
  period_extra = "2019"
)

simulate_scenario_change_in_settlement(
  seed,
  n = 125000,
  prob.Type = c(0.6, 0.25, 0.15),
  prob.Hidden = c(0.35, 0.45, 0.2),
  max_rep_delay = 2,
  max_set_delay = 20,
  settlement_change = "2017"
)
}
\arguments{
\item{seed}{Integer. Set the seed for the random number generators.}

\item{n}{Integer. The number of claims (occurrence dates) to generate.}

\item{prob.Type}{Optional. Numeric vector of size 3 representing the probability distribution of the claim type.}

\item{prob.Hidden}{Optional. Numeric vector of length 3 representing the probability distribution of the unobserved covariate `hidden`.}

\item{max_rep_delay}{Optional. The assumed maximal reporting delay in years.}

\item{max_set_delay}{Optional. The assumed maximal settlement delay in years.}

\item{inflation}{Optional. Numeric vector of size 3 representing the yearly change in the probability distribution of the claim types in the scenario describing the change in claim mix.}

\item{period_extra}{Optional character. Year in which the extreme event takes place. The extreme event lasts for one month and occurs somewhere randomly within the chosen year.}

\item{settlement_change}{Optional character. Year from when a change in settlement delay takes place.}
}
\value{
A portfolio generated along the chosen scenario.
}
\description{
Simulate portfolios along different scenarios using a simulation engine. Four scenarios are currently implemented: a baseline, change in claim mix, extreme event or change in settlement scenario.
}
