% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{split_data}
\alias{split_data}
\title{Splits the data in a hierarchical reserving model into a training, validation and test data set}
\usage{
split_data(obj, observed = NULL, validation = NULL, cv_fold = NULL)
}
\arguments{
\item{obj}{The hierarchical reserving model}

\item{observed}{This input can be \itemize{
  \item A data set to be used for training and validation
  \item A function that extracts the observed data from the full data set
}}

\item{validation}{This input can be \itemize{
  \item A data set to be used for validation
  \item A function that extracts the validation data set from the observed data set
  \item A number between zero and one, which expresses the fraction of the data that will be used for validation
}}

\item{cv_fold}{Number of folds to be used in cross validation}
}
\description{
Splits the data in a hierarchical reserving model into a training, validation and test data set
}
