% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainLadder.R
\name{chainLadder_open}
\alias{chainLadder_open}
\title{Special version of the chain-ladder model for the incremental run-off triangle of the number of open claims.}
\usage{
chainLadder_open(triangle)
}
\arguments{
\item{triangle}{The incremental run-off triangle of the number of open claims.}
}
\description{
This function applies a special version of the chain-ladder model for the incremental run-off triangle of the number of open claims.
}
